+++ @types/express/index.d.ts
@@ -2,8 +2,12 @@ import * as express from 'express'
 
 export declare global {
     namespace Express {
-      interface Request {
+     export interface Request {
         userEmail: string 
+      }
+
+      export interface Request{
+        userId: any
       }
     }
 }
